audio tests, serial mode support
