+++ bigbluebutton-tests/playwright/audio/audio.spec.js
@@ -1,14 +1,14 @@
 const { test } = require('@playwright/test');
 const { testSpeakerButton } = require('../core/elements');
-const { MultiUsers } = require('../user/multiusers');
 const { Audio } = require('./audio');
 
-test.describe.serial('Audio', () => {
+test.describe.parallel('Audio', () => {
   const audio = new Audio();
   test.beforeAll(async ({ browser }) => {
     const context = await browser.newContext();
     const page = await context.newPage();
     await audio.initModPage(page, false);
+    await audio.initUserPage(false, context);
   });
 
   // https://docs.bigbluebutton.org/2.6/release-tests.html#listen-only-mode-automated
@@ -30,39 +30,19 @@ test.describe.serial('Audio', () => {
     await audio.changeAudioInput();
   });
 
+  // https://docs.bigbluebutton.org/2.6/release-tests.html#choosing-different-sources
   test('Keep the last mute state after rejoining audio @ci', async () => {
     await audio.keepMuteStateOnRejoin();
   });
 
+  // Talking Indicator
+  // https://docs.bigbluebutton.org/2.6/release-tests.html#talking-indicator
   test('Mute yourself by clicking the talking indicator', async () => {
-    await audio.keepMuteStateOnRejoin();
+    await audio.muteYourselfByTalkingIndicator();
   });
 
-  /*
-
-  // https://docs.bigbluebutton.org/2.6/release-tests.html#choosing-different-sources
-  test('Keep the last mute state after rejoining audio @ci', async ({ browser, page }) => {
-    const audio = new Audio(browser, page);
-    await audio.init(true, false);
-    await audio.keepMuteStateOnRejoin();
+  // https://docs.bigbluebutton.org/2.6/release-tests.html#talking-indicator
+  test('Mute another user by clicking the talking indicator', async () => {
+    await audio.muteAnotherUser();
   });
-
-  test.describe.parallel('Talking indicator @ci', () => {
-    // https://docs.bigbluebutton.org/2.6/release-tests.html#talking-indicator
-    test('Mute yourself by clicking the talking indicator', async ({ browser, page }) => {
-      const audio = new Audio(browser, page);
-      await audio.init(true, false);
-      await audio.muteYourselfByTalkingIndicator();
-    });
-
-    // https://docs.bigbluebutton.org/2.6/release-tests.html#talking-indicator
-    test('Mute another user by clicking the talking indicator', async ({ browser, context, page }) => {
-      const audio = new MultiUsers(browser, context);
-      await audio.initModPage(page);
-      await audio.initUserPage(false);
-      await audio.muteAnotherUser();
-    });
-  });
-
-  */
 });
