+++ bigbluebutton-tests/playwright/core/media/.~lock.smartSlidesPresentation.pdf#
@@ -1,0 +1,1 @@
+,nt-imdt,imdt,19.05.2023 11:17,file:///home/nt-imdt/.config/libreoffice/4;
